Module oak_remote_attestation::handler

Structs

AsyncEncryptionHandler

Wraps a closure to an underlying function with request encryption and response decryption logic, based on the provided encryption key. AsyncEncryptionHandler can be used when an [AsyncRecipientContextGenerator] is needed.

EncryptionHandler

Wraps a closure to an underlying function with request encryption and response decryption logic, based on the provided encryption key.

PublicKeyInfo

Information about a public key.