Trait oak_remote_attestation::handshaker::AttestationGenerator

pub trait AttestationGenerator: Send + Sync {
    fn generate_attestation(&self, attested_data: &[u8]) -> Result<Vec<u8>>;
}

A trait implementing the functionality of generating a remote attestation report.

An implementation of this trait is expected to run in a TEE (i.e. it is usually in the server).

Required Methods

fn generate_attestation(&self, attested_data: &[u8]) -> Result<Vec<u8>>

Generate a remote attestation report, ensuring that attested_data is cryptographically bound to the result (e.g. via a signature).

That is usually verified by AttestationVerifier::verify_attestation.