Struct oak_remote_attestation::message::ServerIdentity

pub struct ServerIdentity {
    pub version: u8,
    pub ephemeral_public_key: [u8; 32],
    pub random: [u8; 32],
    pub transcript_signature: [u8; 64],
    pub signing_public_key: [u8; 65],
    pub attestation_info: Vec<u8>,
    pub additional_info: Vec<u8>,
    // some fields omitted
}

Server identity message containing remote attestation information and a public key for Diffie-Hellman key negotiation.

Fields
version: u8

Remote attestation protocol version.

ephemeral_public_key: [u8; 32]

Public key needed to establish a session key.

random: [u8; 32]

Random vector sent in messages for preventing replay attacks.

transcript_signature: [u8; 64]

Signature of the SHA-256 hash of all previously sent and received messages. Transcript signature is sent in messages to prevent replay attacks.

Signature must be an IEEE-P1363 encoded ECDSA-P256 signature. https://datatracker.ietf.org/doc/html/rfc6979 https://standards.ieee.org/standard/1363-2000.html

signing_public_key: [u8; 65]

Public key used to sign transcripts.

Public key must be an OpenSSL ECDSA-P256 key, which is represented as 0x04 | X: 32-byte | Y: 32-byte. Where X and Y are big-endian coordinates of an Elliptic Curve point. https://datatracker.ietf.org/doc/html/rfc6979

attestation_info: Vec<u8>

Information used for remote attestation such as a TEE report and a TEE provider’s certificate. TEE report contains a hash of the signing_public_key and additional_info.

Attestation info must be a serialized oak.remote_attestation.AttestationInfo Protobuf message.

additional_info: Vec<u8>

Additional info to be checked when verifying the identity. This may include server configuration details, and inclusion proofs on a verifiable log (e.g., LogEntry on Rekor). The server and the client must be able to agree on a canonical representation of the content to be able to deterministically compute the hash of this field.

Implementations

impl ServerIdentity

pub fn new(
ephemeral_public_key: [u8; 32],
random: [u8; 32],
signing_public_key: [u8; 65],
attestation_info: Vec<u8>,
additional_info: Vec<u8>
) -> Self

pub fn clear_transcript_signature(&mut self)

pub fn set_transcript_signature(&mut self, transcript_signature: &[u8; 64])