Trait oak_remote_attestation::attester::AttestationReportGenerator

pub trait AttestationReportGenerator: Send + Sync {
    // Required method
    fn generate_attestation_report(
        &self,
        attested_data: &[u8]
    ) -> Result<Vec<u8>>;
}

A trait implementing the functionality of generating a remote attestation report.

An implementation of this trait is expected to run in a TEE (i.e. it is usually in the server).

Required Methods

fn generate_attestation_report(&self, attested_data: &[u8]) -> Result<Vec<u8>>

Generate a remote attestation report, ensuring that attested_data is cryptographically bound to the result (e.g. via a signature).